# backend/internal/adapters/httpapi/httpapi.gen.go
//go:build go1.22

// Package httpapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package httpapi

import (
	"errors"
	"fmt"
	"net/http"
	"time"

	"github.com/oapi-codegen/runtime"
)

// Defines values for Indicator.
const (
	IndicatorDegraded    Indicator = "degraded"
	IndicatorOperational Indicator = "operational"
	IndicatorOutage      Indicator = "outage"
	IndicatorUnknown     Indicator = "unknown"
)

// Valid indicates whether the value is a known member of the Indicator enum.
func (e Indicator) Valid() bool {
	switch e {
	case IndicatorDegraded:
		return true
	case IndicatorOperational:
		return true
	case IndicatorOutage:
		return true
	case IndicatorUnknown:
		return true
	default:
		return false
	}
}

// Defines values for Status.
const (
	StatusIdentified    Status = "identified"
	StatusInvestigating Status = "investigating"
	StatusMaintenance   Status = "maintenance"
	StatusMonitoring    Status = "monitoring"
	StatusResolved      Status = "resolved"
	StatusUnknown       Status = "unknown"
)

// Valid indicates whether the value is a known member of the Status enum.
func (e Status) Valid() bool {
	switch e {
	case StatusIdentified:
		return true
	case StatusInvestigating:
		return true
	case StatusMaintenance:
		return true
	case StatusMonitoring:
		return true
	case StatusResolved:
		return true
	case StatusUnknown:
		return true
	default:
		return false
	}
}

// Error defines model for Error.
type Error struct {
	// Code Stable, machine-readable error code.
	//
	// Example: feed_unreachable
	Code string `json:"code"`

	// Message Human-readable explanation.
	Message string `json:"message"`
}

// Feed defines model for Feed.
type Feed struct {
	CreatedAt time.Time `json:"created_at"`

	// CurrentStatus Derived from the most recent incident; null when there are no items.
	CurrentStatus *Status `json:"current_status,omitempty"`

	// Enabled When false, the feed is not polled.
	Enabled bool `json:"enabled"`

	// GroupId 0 == Ungrouped.
	GroupId int64 `json:"group_id"`
	Id      int64 `json:"id"`

	// LastError Last failure message; null when healthy.
	LastError *string `json:"last_error,omitempty"`

	// LastFetchedAt Last poll attempt (RFC3339 UTC); null until the first poll.
	LastFetchedAt *time.Time `json:"last_fetched_at,omitempty"`

	// LastSuccessAt Last successful poll (RFC3339 UTC); null until the first success.
	LastSuccessAt *time.Time `json:"last_success_at,omitempty"`

	// RefreshIntervalSec Poll cadence in seconds.
	RefreshIntervalSec int `json:"refresh_interval_sec"`

	// Title Example: GitHub Status - Incident History
	Title     string    `json:"title"`
	UpdatedAt time.Time `json:"updated_at"`

	// Url Example: https://www.githubstatus.com/history.atom
	Url string `json:"url"`
}

// Indicator The traffic-light shown next to a system, derived server-side from its most recent incident: operational (green), degraded (yellow), outage (red), or unknown (grey — never polled, or the status could not be derived).
type Indicator string

// Status Normalized, best-effort incident state derived from the feed content. May be "unknown" for feeds that don't follow the Atlassian Statuspage convention.
type Status string

// StatusItem defines model for StatusItem.
type StatusItem struct {
	// ContentHtml Raw entry body (all updates), preserved verbatim.
	ContentHtml *string `json:"content_html,omitempty"`

	// ContentText Stripped plaintext, for preview/search.
	ContentText *string   `json:"content_text,omitempty"`
	FeedId      int64     `json:"feed_id"`
	FetchedAt   time.Time `json:"fetched_at"`

	// Guid Stable external id (Atom <id> / RSS <guid> / fallback link).
	Guid string `json:"guid"`
	Id   int64  `json:"id"`

	// Link Permalink to the incident.
	Link *string `json:"link,omitempty"`

	// PublishedAt When the provider published the entry; null for feeds that ship no timestamps (fetched_at is then the only thing we know).
	PublishedAt *time.Time `json:"published_at,omitempty"`

	// Status Normalized, best-effort incident state derived from the feed content. May be "unknown" for feeds that don't follow the Atlassian Statuspage convention.
	Status Status `json:"status"`
	Title  string `json:"title"`

	// UpdatedAt When the provider last revised the entry; null when it never said.
	UpdatedAt *time.Time `json:"updated_at,omitempty"`
}

// SubscribeFeedRequest defines model for SubscribeFeedRequest.
type SubscribeFeedRequest struct {
	// Enabled Whether the feed is polled; omitted means true.
	Enabled *bool `json:"enabled,omitempty"`

	// GroupId 0 == Ungrouped, which is also what an omitted value falls back to. Must reference an existing group.
	GroupId *int64 `json:"group_id,omitempty"`

	// RefreshIntervalSec Poll cadence in seconds. Any positive value is accepted; omitted means 300. Note that the scheduler only looks for due feeds every few seconds, so a cadence below that tick is effectively rounded up to it.
	RefreshIntervalSec *int `json:"refresh_interval_sec,omitempty"`

	// Title Display name for the system. Required, trimmed, and unique case-insensitively across feeds.
	Title string `json:"title"`

	// Url The RSS/Atom endpoint to subscribe to. Must be an http(s) URL, and is fetched once before the feed is stored — an endpoint that cannot be read or parsed is rejected with 422 and nothing is saved.
	//
	// Example: https://www.githubstatus.com/history.atom
	Url string `json:"url"`
}

// SystemOverview One system as the main page needs it: the feed itself plus the derived indicator. Incidents are intentionally absent — see /feeds/{feedId}/items.
type SystemOverview struct {
	Feed Feed `json:"feed"`

	// Indicator The traffic-light shown next to a system, derived server-side from its most recent incident: operational (green), degraded (yellow), outage (red), or unknown (grey — never polled, or the status could not be derived).
	Indicator Indicator `json:"indicator"`

	// LastUpdatedAt Newest incident's timestamp; null when the system has no incidents.
	LastUpdatedAt *time.Time `json:"last_updated_at,omitempty"`
}

// FeedId defines model for FeedId.
type FeedId = int64

// GroupId defines model for GroupId.
type GroupId = int64

// BadRequest defines model for BadRequest.
type BadRequest = Error

// Conflict defines model for Conflict.
type Conflict = Error

// NotFound defines model for NotFound.
type NotFound = Error

// UnprocessableEntity defines model for UnprocessableEntity.
type UnprocessableEntity = Error

// ListFeedItemsParams defines parameters for ListFeedItems.
type ListFeedItemsParams struct {
	// Since Only items published at or after this RFC3339 timestamp.
	Since *time.Time `form:"since,omitempty" json:"since,omitempty"`

	// Limit Maximum number of items to return.
	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`
}

// SubscribeFeedJSONRequestBody defines body for SubscribeFeed for application/json ContentType.
type SubscribeFeedJSONRequestBody = SubscribeFeedRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// SubscribeFeed Subscribe to a new feed.
	// (POST /feeds)
	SubscribeFeed(w http.ResponseWriter, r *http.Request)
	// ListFeedItems List a feed's status items (incidents), newest first.
	// (GET /feeds/{feedId}/items)
	ListFeedItems(w http.ResponseWriter, r *http.Request, feedId FeedId, params ListFeedItemsParams)
	// GetOverview Latest known status of every subscribed system.
	// (GET /overview)
	GetOverview(w http.ResponseWriter, r *http.Request)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// SubscribeFeed operation middleware
func (siw *ServerInterfaceWrapper) SubscribeFeed(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.SubscribeFeed(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListFeedItems operation middleware
func (siw *ServerInterfaceWrapper) ListFeedItems(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "feedId" -------------
	var feedId FeedId

	err = runtime.BindStyledParameterWithOptions("simple", "feedId", r.PathValue("feedId"), &feedId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "int64", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "feedId", Err: err})
		return
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params ListFeedItemsParams

	// ------------- Optional query parameter "since" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "since", r.URL.Query(), &params.Since, runtime.BindQueryParameterOptions{Type: "string", Format: "date-time"})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "since"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "since", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "limit", r.URL.Query(), &params.Limit, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "limit"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "limit", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListFeedItems(w, r, feedId, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetOverview operation middleware
func (siw *ServerInterfaceWrapper) GetOverview(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetOverview(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/overview", wrapper.GetOverview)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/feeds", wrapper.SubscribeFeed)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/feeds/{feedId}/items", wrapper.ListFeedItems)

	return m
}